+++ CHANGELOG.md
@@ -8,6 +8,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Fixed
+
+- Fixed bug where root game tags had an empty HierarchyTags.
+
 ## [0.1.0-beta.3]
 
 ### Added